fix float conversion
[git-p4: depot-paths = "//depot/qt/main/": change = 151592]

+++ src/gui/painting/qmatrix.cpp
@@ -345,10 +345,10 @@ QRectF QMatrix::mapRect(const QRectF &rect) const
 {
     QRectF result;
     if (_m12 == 0.0F && _m21 == 0.0F) {
-        int x = int(_m11*rect.x() + _dx);
-        int y = int(_m22*rect.y() + _dy);
-        int w = int(_m11*rect.width());
-        int h = int(_m22*rect.height());
+        float x = _m11*rect.x() + _dx;
+        float y = _m22*rect.y() + _dy;
+        float w = _m11*rect.width();
+        float h = _m22*rect.height();
         if (w < 0) {
             w = -w;
             x -= w;
